# docs/swagger.yaml
definitions:
  fail.Response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/recipes/{recipe_id}/categories:
    put:
      consumes:
      - application/json
      description: Set recipe categories
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Set recipe categories
      tags:
      - recipe
  /v1/recipes/{recipe_id}/favourite:
    delete:
      consumes:
      - application/json
      description: Remove recipe from favourite
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove recipe from favourite
      tags:
      - recipe
    post:
      consumes:
      - application/json
      description: Add recipe to favourite
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Add recipe to favourite
      tags:
      - recipe
  /v1/recipes/{recipe_id}/rate:
    post:
      consumes:
      - application/json
      description: Rate recipe
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Rate recipe
      tags:
      - recipe
  /v1/recipes/{recipe_id}/save:
    post:
      consumes:
      - application/json
      description: Save recipe to user's recipe book
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Save recipe
      tags:
      - recipe
...
  /v1/recipes{recipe_id}/save:
    delete:
      consumes:
      - application/json
      description: Remove recipe from recipe book
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mephistolie_chefbook-backend-api-gateway_internal_transport_http_helpers_response.MessageBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fail.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fail.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove recipe from recipe book
      tags:
      - recipe
...
swagger: "2.0"